APIs/Oauth/hs-new/requirements.txt
aiofiles==0.7.0
certifi==2021.5.30
httptools==0.2.0
hubspot-api-client==3.8.2
multidict==5.1.0
python-dateutil==2.8.2
sanic==21.6.2
Sanic-Cors==1.0.0
sanic-plugin-toolkit==1.0.1
sanic-routing==0.7.1
six==1.16.0
ujson==4.0.2
urllib3==1.26.6
uvloop==0.15.3
websockets==9.1
